Clean up ledgerWallet.ts into seperate files
