+++ app/lib/web3/Web3Adapter.ts
@@ -4,7 +4,7 @@ import * as Web3 from "web3";
 import { makeEthereumAddressChecksummed } from "../../modules/web3/utils";
 import { EthereumAddress, EthereumAddressWithChecksum, EthereumNetworkId } from "../../types";
 import { delay } from "../../utils/delay";
-import { promisify } from "../../utils/promisify";
+import { promisify } from "bluebird";
 
 class Web3Error extends Error {}
 export class NotEnoughEtherForGasError extends Error {}
