+++ app/lib/web3/ledger-wallet/types.ts
@@ -1,0 +1,10 @@
+
+export interface ILedgerConfig {
+    version: string;
+    arbitraryDataEnabled: 0 | 1;
+  }
+  
+  export interface IDerivationPathToAddress {
+    [derivationPath: string]: string;
+  }
+
