+++ app/modules/access-wallet/errors.ts
@@ -5,7 +5,7 @@ import {
 import { createMessage, TMessage } from "../../components/translatedMessages/utils";
 import { BrowserWalletError } from "../../lib/web3/BrowserWallet";
 import { mapSignerErrorToErrorMessage } from "../../lib/web3/errors";
-import { LedgerError } from "../../lib/web3/LedgerWallet";
+import { LedgerError } from "../../lib/web3/ledger-wallet/LedgerWallet";
 import { LightWalletError } from "../../lib/web3/LightWallet";
 import { SignerError } from "../../lib/web3/Web3Manager";
 import { EthereumAddress } from "../../types";
