+++ app/modules/wallet-selector/ledger-wizard/errors.ts
@@ -1,6 +1,6 @@
 import { LedgerErrorMessage } from "../../../components/translatedMessages/messages";
 import { createMessage, TMessage } from "../../../components/translatedMessages/utils";
-import { LedgerContractsDisabledError, LedgerLockedError } from "../../../lib/web3/LedgerWallet";
+import { LedgerContractsDisabledError, LedgerLockedError } from "../../../lib/web3/ledger-wallet/LedgerWallet";
 
 export function mapLedgerErrorToErrorMessage(error: Error): TMessage {
   let messageType = LedgerErrorMessage.GENERIC_ERROR;
