+++ package.json
@@ -198,6 +198,7 @@
     "@ledgerhq/web3-subprovider": "^4.41.1",
     "@sentry/browser": "^4.3.0",
     "bignumber.js": "^5.0.0",
+    "bluebird": "^3.5.3",
     "bootstrap": "^4.1.3",
     "chart.js": "^2.7.2",
     "classnames": "^2.2.6",
@@ -255,7 +256,7 @@
     "tweetnacl": "^1.0.0",
     "tweetnacl-util": "^0.15.0",
     "web3": "0.20.2",
-    "web3-provider-engine": "^13.5.0",
+    "web3-provider-engine": "^14.1.0",
     "web3-utils": "^1.0.0-beta.35",
     "yup": "^0.26.7"
   }
